requirements.txt
streamlit
langchain_community
langchain_groq
langchain_core
python-dotenv